Refactor - extract function to utils
