+++ playground/18_frustum_picking/index.ts
@@ -5,9 +5,9 @@ import {
     Camera,
     Vec2, vec2, ZERO2,
     vec3, mul3,
-    Mat4, mat4, apply4x4, clone4x4, frustum4x4,
+    mat4,
     Color, color, colors,
-    uint2bytes, makeEventCoordsGetter, spherical2zxy, deg2rad,
+    uint2bytes, makeEventCoordsGetter, spherical2zxy, deg2rad, makePixelViewProjMat,
 } from 'lib';
 import { observable, computed } from 'util/observable';
 import { createControls } from 'util/controls';
@@ -98,11 +98,13 @@ function renderFrame(state: State): void {
     renderScene(state, pixelIdx);
 }
 
+const _transformMat = mat4();
 function findCurrentPixel({ runtime, framebuffer, camera, objects, idProgram, pixelCoord }: State): number {
     runtime.setRenderTarget(framebuffer);
     runtime.setClearColor(colors.NONE);
     runtime.clearBuffer('color|depth');
-    const transformMat = makeIdViewProjMat(camera, pixelCoord);
+    const transformMat = _transformMat;
+    makePixelViewProjMat(camera, pixelCoord, transformMat);
     idProgram.setUniform('u_view_proj', transformMat);
     for (const { primitive, modelMat, id } of objects) {
         primitive.setProgram(idProgram);
@@ -152,35 +154,4 @@ function makeObjects(runtime: Runtime): {
         ),
     ];
     return { objects, program, idProgram };
-}
-
-// Makes view-projection matrix that renders only one pixel.
-// View matrix is the same with default scene view matrix.
-// Default perspective projection matrix is replaced with frustum projection matrix that
-// "covers" only one required pixel
-const _idViewProjMat = mat4();
-function makeIdViewProjMat(camera: Camera, { x, y }: Vec2): Mat4 {
-    // Calculate frustum from perspective parameters.
-    // |top| = |bottom| = zNear * tan(fov / 2)
-    // |left| = |right| = aspect * |top|
-    const dy = camera.getZNear() * Math.tan(camera.getYFov() / 2);
-    const dx = camera.getAspect() * dy;
-    const { x: xViewport, y: yViewport } = camera.getViewportSize();
-    // Full [left, right] * [bottom, top] range corresponds to [0, viewport_width] * [0, viewport_height] screen.
-    // [0, W] -> [-dx, +dx] => x -> dx * (x * 2 / W - 1)
-    // [0, H] -> [-dy, +dy] => y -> dy * (y * 2 / H - 1)
-    // Select part that corresponds to a specific (x, y) pixel.
-    // In an arbitrary n pixels range [0, n] i-th (0 <= i < n) pixel occupies [i, i + 1] range.
-    const x1 = dx * (2 * x / xViewport - 1);
-    const x2 = dx * (2 * (x + 1) / xViewport - 1);
-    const y1 = dy * (2 * y / yViewport - 1);
-    const y2 = dy * (2 * (y + 1) / yViewport - 1);
-
-    const mat = _idViewProjMat;
-    clone4x4(camera.getViewMat(), mat);
-    apply4x4(mat, frustum4x4, {
-        left: x1, right: x2, bottom: y1, top: y2,
-        zNear: camera.getZNear(), zFar: camera.getZFar(),
-    });
-    return mat;
 }
